+++ Test/Cases/d10_for_typed.pas
@@ -1,0 +1,13 @@
+unit d10_for_typed;
+
+interface
+
+implementation
+
+procedure Demo;
+begin
+  for var L: Integer := 0 to 9 do
+    WriteLn(L);
+end;
+
+end.